refactor: 컴포넌트 기능 별 디렉토리 추가
